requirements.txt
streamlit
numpy
pandas
matplotlib
torch
torchvision
Pillow
gdown
scikit-learn==1.3.0
joblib
